fix(desktop): keep renderer tool-loop parser browser-safe
