+++ packages/maker-core/src/types/events.ts
@@ -11,7 +11,17 @@
 
 import type { WorkflowProgressEntry } from '@cindy/maker-shared/agent-task';
 import type { SubagentObservation } from '@cindy/maker-shared/subagent-observation';
+import {
+  parseToolLoopErrorDetails,
+  type ToolLoopErrorDetails,
+} from '@cindy/maker-shared/tool-loop-error';
 import type { PiRuntimeCapabilityManifest } from './pi-runtime-capabilities.js';
+
+export {
+  parseToolLoopErrorDetails,
+  type ToolLoopErrorDetails,
+  type ToolLoopErrorKind,
+} from '@cindy/maker-shared/tool-loop-error';
 
 export type AgentEventType =
   | 'text'                  // 流式文本输出（增量或完整）
@@ -41,38 +51,6 @@ export interface TurnDiffEventData {
   cwd: string;
   /** False when concurrent provider snapshots could not be safely composed. */
   isComplete?: boolean;
-}
-
-/** Stable, renderer-safe details for a tool-loop guard terminal error. */
-export type ToolLoopErrorKind = 'consecutive' | 'pingpong' | 'rotation' | 'contract';
-
-export interface ToolLoopErrorDetails {
-  kind: ToolLoopErrorKind;
-  count: number;
-}
-
-/**
- * Accept only the bounded, enum-shaped details emitted by maker-core. This is
- * shared by main persistence and renderer hydration so untrusted event data
- * cannot become arbitrary UI interpolation or durable content.
- */
-export function parseToolLoopErrorDetails(value: unknown): ToolLoopErrorDetails | undefined {
-  if (!value || typeof value !== 'object') return undefined;
-  const raw = value as { kind?: unknown; count?: unknown };
-  const kind = raw.kind;
-  const count = raw.count;
-  if (
-    kind !== 'consecutive' &&
-    kind !== 'pingpong' &&
-    kind !== 'rotation' &&
-    kind !== 'contract'
-  ) {
-    return undefined;
-  }
-  if (typeof count !== 'number' || !Number.isSafeInteger(count) || count < 1 || count > 100_000) {
-    return undefined;
-  }
-  return { kind, count };
 }
 
 export interface AgentErrorEventData {
