+++ packages/maker-shared/package.json
@@ -68,6 +68,7 @@
     "./turn-continuation": "./src/turnContinuation.ts",
     "./tool-use-descriptor": "./src/toolUseDescriptor.ts",
     "./tool-result-compaction": "./src/toolResultCompaction.ts",
+    "./tool-loop-error": "./src/toolLoopError.ts",
     "./update-channel": "./src/updateChannel.ts",
     "./url-text-boundary": "./src/urlTextBoundary.ts",
     "./usage-format": "./src/usageFormat.ts",
