+++ packages/maker-shared/src/index.ts
@@ -50,6 +50,7 @@ export * from './subagentWorkspace.js';
 export * from './systemCard.js';
 export * from './toolUseDescriptor.js';
 export * from './toolResultCompaction.js';
+export * from './toolLoopError.js';
 export * from './thinkingText.js';
 export * from './turnContinuation.js';
 export * from './updateChannel.js';
